Simplifying useState for status notifications for upload/fetch/delete functions; combined exports from utils using a single index.js; refactored runNotification further
