+++ src/utils/index.js
@@ -1,0 +1,17 @@
+import { docTypes } from "./form-helper";
+import { runNotification } from "./notification-helper";
+import {
+  SOLID_IDENTITY_PROVIDER,
+  handleFiles,
+  fetchDocuments,
+  deleteDocuments,
+} from "./session-helper";
+
+export {
+  docTypes,
+  runNotification,
+  SOLID_IDENTITY_PROVIDER,
+  handleFiles,
+  fetchDocuments,
+  deleteDocuments,
+};
