+++ src/utils/notification-helper.js
@@ -5,21 +5,14 @@ export const runNotification = (
   setStatusFunc,
   setTimeoutFunc
 ) => {
-  // if local timeoutID exist, clear previous message and
-  // run new status notification message
+  // if local timeoutID exist, clear previous message and run new status notification message
   if (timeoutID) {
     clearTimeout(timeoutID);
-    setStatusFunc({ state: true, message });
   }
-
+  setStatusFunc(message);
   // set timeout for potentially new notification
   const timeout = setTimeout(() => {
-    setStatusFunc({
-      state: false,
-      message: "",
-    });
+    setStatusFunc("");
   }, time * 1000);
-
   setTimeoutFunc(timeout);
-  setStatusFunc({ state: true, message });
 };
